requirements.txt
addict==2.4.0
basicsr==1.4.2
cityscapesscripts==2.2.2
einops==0.7.0
lmdb==1.4.1
matplotlib==3.1.2
mc==0.0
numpy==1.17.4
onnx==1.15.0
opencv_contrib_python==4.3.0.36
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
packaging==20.3
parrots==1.0.3
Pillow==7.0.0
Pillow==10.2.0
prettytable==3.10.0
pytorch_lightning==1.5.0
PyYAML==6.0.1
PyYAML==6.0.1
regex==2023.12.25
scikit_image==0.16.2
scipy==1.3.3
share==1.0.4
tensorboardX==2.6.2.2
tifffile==2024.2.12
timm==0.6.12
torch==2.0.1
torch_tb_profiler==0.4.3
torchvision==0.15.2
turbojpeg==0.0.2
yapf==0.32.0